requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.40.0.post2
bleach==6.0.0
blis==0.7.9
brotlipy==0.7.0
catalogue==2.0.8
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
confection==0.0.4
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1686613057838/work
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.9
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
executing==1.2.0
faiss-gpu==1.7.2
fastjsonschema==2.17.1
filelock==3.12.2
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
gradio_client==0.6.1
greenlet==2.0.2
h11==0.14.0
h5py==3.11.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.23.0
idna @ file:///croot/idna_1666125576474/work
install==1.3.5
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langchain==0.0.229
langchainplus-sdk==0.0.20
langcodes==3.3.0
loguru==0.7.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
# Editable install with no version control (megatron_core==0.4.0rc0)
-e /mnt
mistune==2.0.5
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx @ file:///croot/networkx_1678964333703/work
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.3
plotly==5.16.1
preshed==3.0.8
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-readability-metrics==1.4.5
pyarrow==12.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.9
Pygments==2.15.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.6.3
replicate==0.15.3
requests @ file:///croot/requests_1682607517574/work
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.18
srsly==2.4.6
stack-data==0.6.2
sympy @ file:///croot/sympy_1668202399572/work
tenacity==8.2.2
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.19.1
torch==2.1.0
torchaudio==2.3.0
torchvision==0.16.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformer-engine @ git+https://github.com/NVIDIA/TransformerEngine.git@9709147e35ec09062d29712042c3aba3cb1ab57e
transformers==4.41.1
triton==2.1.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1686163155763/work
wandb==0.15.4
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
websockets==11.0.3
widgetsnbextension==4.0.7
wordcloud==1.9.2
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
